refactor: 스타일 모듈 제거
컨테이너 제외한 버튼을 variant로 설정하는 방식으로 변경되어 분리할 필요성이 사라짐
